Refactor Buttons
Moves common button methods to a module and unique
elements to independent classes.

Hold and Recall buttons remain so tightly coupled that they share a
class even in this refactor.
